Reorganize MCP ITs a bit to make them easier to split
