+++ langchain4j-mcp/src/test/java/dev/langchain4j/mcp/client/integration/McpToolsHttpTransportIT.java
@@ -1,5 +1,7 @@
 package dev.langchain4j.mcp.client.integration;
 
+import static dev.langchain4j.mcp.client.integration.McpServerHelper.skipTestsIfJbangNotAvailable;
+import static dev.langchain4j.mcp.client.integration.McpServerHelper.startServerHttp;
 import static org.assertj.core.api.Assertions.fail;
 
 import dev.langchain4j.mcp.client.DefaultMcpClient;
@@ -7,33 +9,23 @@ import dev.langchain4j.mcp.client.McpClient;
 import dev.langchain4j.mcp.client.transport.McpTransport;
 import dev.langchain4j.mcp.client.transport.http.HttpMcpTransport;
 import java.io.IOException;
-import java.net.URISyntaxException;
 import java.time.Duration;
-import java.util.Arrays;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class McpHttpTransportIT extends McpTransportTestBase {
+public class McpToolsHttpTransportIT extends McpToolsTestBase {
 
-    private static final Logger log = LoggerFactory.getLogger(McpHttpTransportIT.class);
+    private static final Logger log = LoggerFactory.getLogger(McpToolsHttpTransportIT.class);
     private static Process process;
 
     @BeforeAll
-    public static void setup() throws IOException, InterruptedException, TimeoutException, URISyntaxException {
+    public static void setup() throws IOException, InterruptedException, TimeoutException {
         skipTestsIfJbangNotAvailable();
-        String path = getPathToScript("tools_mcp_server.java");
-        String[] command = new String[] {getJBangCommand(), "--quiet", "--fresh", "run", path};
-        log.info("Starting the MCP server using command: " + Arrays.toString(command));
-        process = new ProcessBuilder().command(command).inheritIO().start();
-        waitForPort(8080, 120);
-        log.info("MCP server has started");
+        process = startServerHttp("tools_mcp_server.java");
         McpTransport transport = new HttpMcpTransport.Builder()
                 .sseUrl("http://localhost:8080/mcp/sse")
                 .logRequests(true)
@@ -43,21 +35,6 @@ public class McpHttpTransportIT extends McpTransportTestBase {
                 .transport(transport)
                 .toolExecutionTimeout(Duration.ofSeconds(4))
                 .build();
-    }
-
-    private static void waitForPort(int port, int timeoutSeconds) throws InterruptedException, TimeoutException {
-        Request request = new Request.Builder().url("http://localhost:" + port).build();
-        long start = System.currentTimeMillis();
-        OkHttpClient client = new OkHttpClient();
-        while (System.currentTimeMillis() - start < timeoutSeconds * 1000) {
-            try {
-                client.newCall(request).execute();
-                return;
-            } catch (IOException e) {
-                TimeUnit.SECONDS.sleep(1);
-            }
-        }
-        throw new TimeoutException("Port " + port + " did not open within " + timeoutSeconds + " seconds");
     }
 
     @AfterAll
