+++ langchain4j-mcp/src/test/java/dev/langchain4j/mcp/client/integration/McpToolsStdioTransportIT.java
@@ -1,23 +1,23 @@
 package dev.langchain4j.mcp.client.integration;
 
+import static dev.langchain4j.mcp.client.integration.McpServerHelper.getJBangCommand;
+import static dev.langchain4j.mcp.client.integration.McpServerHelper.getPathToScript;
+
 import dev.langchain4j.mcp.client.DefaultMcpClient;
 import dev.langchain4j.mcp.client.transport.McpTransport;
 import dev.langchain4j.mcp.client.transport.stdio.StdioMcpTransport;
-import java.net.URISyntaxException;
 import java.time.Duration;
 import java.util.List;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 
-public class McpStdioTransportIT extends McpTransportTestBase {
+public class McpToolsStdioTransportIT extends McpToolsTestBase {
 
     @BeforeAll
-    public static void setup() throws URISyntaxException {
-        skipTestsIfJbangNotAvailable();
-        String path = getPathToScript("tools_mcp_server.java");
-        System.out.println(path);
+    public static void setup() {
         McpTransport transport = new StdioMcpTransport.Builder()
-                .command(List.of(getJBangCommand(), "--quiet", "--fresh", "run", path))
+                .command(List.of(
+                        getJBangCommand(), "--quiet", "--fresh", "run", getPathToScript("tools_mcp_server.java")))
                 .logEvents(true)
                 .build();
         mcpClient = new DefaultMcpClient.Builder()
